Make formats apply to URL textures
